refactor: team 관련 type, constant 정리
