+++ packages/user/src/constants/team.ts
@@ -1,8 +1,10 @@
-import { Category } from 'src/types/user';
+import type { Category } from 'src/types/category';
 
-export type TeamDescription = { label: string; details: string };
+export const TEAM_TYPES = ['pet', 'place', 'travel', 'leisure'] as const;
 
-export const TEAM_DESCRIPTIONS: Record<Category, TeamDescription> = {
+export type TeamDescriptions = { label: string; details: string };
+
+export const TEAM_DESCRIPTIONS: Record<Category, TeamDescriptions> = {
 	pet: { label: '펫 프렌들리', details: '반려동물의 편안하고\n안전한 여행을 위한' },
 	travel: { label: '여행의 정석', details: '아웃도어 활동을 쉽고 편하게' },
 	place: { label: '공간활용의 기술', details: '많은 물건도 구석구석 알차게' },
