+++ packages/user/src/types/category.d.ts
@@ -1,0 +1,3 @@
+import { TEAM_TYPES } from 'src/constants/team';
+
+export type Category = (typeof TEAM_TYPES)[number];
